fixes pr reviews
