+++ .eslintrc.json
@@ -5,7 +5,7 @@
 		"@typescript-eslint"
 	],
 	"parserOptions": {
-		"project": ["./tsconfig.json"] // Specify it only for TypeScript files,
+		"project": ["./tsconfig.json"]
 	},
 	"extends": [
 		"plugin:@typescript-eslint/recommended"
@@ -43,39 +43,6 @@
 		"no-switch-case-fallthrough": "off",
 		"no-fallthrough": ["error", {"commentPattern": "fall[\\s|\\-]?through"}],
 
-		// proposals
-		"eol-last": "error",
-		"block-spacing": "error",
-		"comma-style": "error",
-		"yoda": "error",
-		"linebreak-style": "error",
-		"key-spacing": "error",
-		"new-parens": "error",
-		"func-call-spacing": "error",
-		"no-lonely-if": "error",
-		"no-multiple-empty-lines": "error",
-		"no-unneeded-ternary": "error",
-		"object-curly-newline": "error",
-		"operator-assignment": "error",
-		"array-bracket-spacing": "error",
-		"arrow-body-style": ["error", "as-needed"],
-		"arrow-spacing": "error",
-		"prefer-const": "error",
-		"prefer-template": "error",
-		"rest-spread-spacing": "error",
-		"template-curly-spacing": "error",
-		"no-trailing-spaces": "error",
-		"no-mixed-operators": "error",
-		"brace-style": "error",
-		"max-lines": ["error", 750],
-		"max-len": ["error", { "code": 160, "tabWidth": 4, "ignoreTemplateLiterals": true}],
-		"max-depth": "error",
-		"max-lines-per-function" : "error",
-		"max-nested-callbacks": "error",
-		"max-params": ["error", {"max": 10}],
-		"require-await": "error",
-		"no-implicit-coercion": ["error", {"boolean":  false}],
-
 		// https://github.com/typescript-eslint/typescript-eslint/tree/master/packages/eslint-plugin/docs/rules
 		"@typescript-eslint/no-explicit-any": "off",
 		"@typescript-eslint/explicit-module-boundary-types": "off",
@@ -91,16 +58,5 @@
 		"@typescript-eslint/no-unused-vars": ["error", {"args":  "none"}],
 		"@typescript-eslint/keyword-spacing": "off",
 		"@typescript-eslint/no-var-requires": "off"
-	},
-	"overrides": [
-		{
-			"files": ["*.ts"], // Your TypeScript files extension,
-			"parserOptions": {
-				"project": ["./tsconfig.json"] // Specify it only for TypeScript files,
-			},
-			"rules": {
-				"@typescript-eslint/explicit-member-accessibility": ["error", {"overrides": {"accessors":  "explicit", "constructors": "off"} }]
-			}
-		}
-	]
+	}
 }
