+++ .github/workflows/pr-test.yml
@@ -24,5 +24,5 @@ jobs:
       - run: npm run test
 
       - name: Run ESLint
-        run: npx eslint ./src
+        run: npx eslint "src/**/*.ts"
         working-directory: ./
