+++ src/locker/redis/redis-locker.ts
@@ -1,5 +1,6 @@
 import type Redlock from 'redlock';
-import { Lock, Locker } from "./locker.interface";
+import {Locker} from "../locker.interface";
+import {RedisLock} from "./redis-lock";
 
 export class RedisLocker implements Locker {
 	public defaultLockDurationMs: number = 500;
@@ -9,13 +10,5 @@ export class RedisLocker implements Locker {
 	public lock(key: string, maxLockDurationMs?: number): Promise<RedisLock> {
 		return this.redLock.lock(`${this.lockNamespace}:${key}`, maxLockDurationMs ?? this.defaultLockDurationMs)
 			.then(lock => new RedisLock(key, lock));
-	}
-}
-
-export class RedisLock implements Lock {
-	constructor(public readonly key: string, private readonly lock: Redlock.Lock) {}
-
-	public unlock(): Promise<void> {
-		return this.lock.unlock();
 	}
 }
