+++ src/no-resolved-value-error.ts
@@ -1,0 +1,5 @@
+export class NoResolvedValueError extends Error {
+	constructor(message: string) {
+		super(message);
+	}
+}
